refactor(report): organise the report UI as atoms, molecules and organisms
The components sat in one flat `components/` folder next to the markup
chunks in `html/`, so nothing said which pieces compose which. The layout
now mirrors the ui-v3 tiers in crocova, which that project documents as
atoms for building blocks, molecules for composites and organisms for
complex components.

`atoms/` holds button, tab-button, icon, pill, legend-item and select.
`molecules/` holds pill-group, legend and select-field, each composing its
atom. `organisms/` holds the eight markup chunks that were in `html/`.
`html.ts` stays the composition root the report builder imports.

Splitting the three composites into an atom and a molecule is what the
tiers are for: `select-field` now pairs a label with the `select` atom
rather than spelling the control out itself, matching how crocova's
InputField wraps its Input.

The emitted report is byte-identical.

Tests stay in `tests/unit/`. crocova co-locates them in `__tests__/` inside
each tier, but vitest here only collects `tests/**/*.test.ts`, so a
co-located test would be silently skipped.
