+++ packages/nestjs-doctor/src/report/ui/organisms/tab-schema.ts
@@ -1,4 +1,4 @@
-import { iconButton } from "../components/button.js";
+import { iconButton } from "../atoms/button.js";
 
 export const TAB_SCHEMA = `
 <!-- ── Tab: Schema ── -->
